requirements.txt
absl-py==2.3.1
torchtune==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
biopython==1.81
bleach==6.2.0
blobfile==3.1.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
datasketch==1.6.5
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
Django==5.1.3
edlib==1.3.9.post1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.13.1
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.6.1
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.10
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
ipython-genutils==0.2.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kagglehub==0.3.13
kiwisolver==1.4.9
lark==1.2.2
lesscpy==0.15.1
Levenshtein==0.27.1
llvmlite==0.44.0
lxml==6.0.1
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mistune==3.1.4
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.4.5
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
openpyxl==3.1.5
overrides==7.7.0
p_tqdm==1.4.2
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parasail==1.3.4
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
ply==3.11
pox==0.3.6
ppft==1.7.7
primer3-py==2.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pycryptodomex==3.23.0
Pygments==2.19.2
pyparsing==3.2.4
pyspoa==0.3.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.14.1
reedsolo==1.7.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typeguard==4.4.4
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xlsxwriter==3.2.8
xxhash==3.5.0
yarl==1.20.1
